more robust curve fitting (#13)
Previously the fitting sometimes does not converge, particularly when the the IC50 is outside the data range. This resulted in a RuntimeError from scipy.optimize.curve_fit.

Now the code has been re-factored to first try the curve fitting with scipy.optimize.curve_fit, and then if that fails try with scipy.minimize (including bounds on key parameters).

In practice, this seems to have fixed the issue with the problematic HIV curves in this issue:
https://github.com/jbloomlab/neutcurve/issues/12

The reason why some optimizers work better on some curves then others it not clear, however.